generation/requirements.txt
accelerate==1.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
canonical==0.52.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ckms==0.65.0
click==8.1.8
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
diffusers @ git+https://github.com/huggingface/diffusers@328e0d20a7b996f9bdb04180457eb08c1b42a76e
dnspython==2.7.0
dsnparse==0.2.1
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==33.3.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
h11==0.14.0
headless==1.12.4
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
inflect==7.5.0
iso3166==2.1.1
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
phonenumbers==8.13.53
pillow==11.1.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
roman==5.0
safetensors==0.5.2
scipy==1.15.1
sentencepiece==0.2.0
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
termcolor==2.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
triton==3.1.0
typeguard==4.4.1
typing_extensions==4.12.2
unimatrix==1.14.0
urllib3==2.3.0
zipp==3.21.0
